TG_Bot/requirement.txt
Python 3.8.10

pip  22.3.1
python-telegram-bot   13.14
requests  2.28.1